requirements.txt
Django==1.5
Pillow==2.2.2
South==0.8.4
argparse==1.2.1
dj-database-url==0.2.2
dj-static==0.0.5
django-filebrowser==3.5.2
django-geoposition==0.1.5
django-grappelli==2.4.8
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
